BE/requirements.txt
langchain
langchain-google-genai
chromadb
pymongo
python-dotenv
fastapi
uvicorn
langgraph